Move PUT tests into separate file
